Refactor quote extraction and timestamp utilities
