+++ apps/web/src/app/api/examples/sales-coaching/analyze/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { AnalysisResult, InsightItem } from "@/types/sales-coaching";
-import { parseSrt } from "./utils";
+import { extractTimestampFromContent, parseSrt, truncateQuote } from "./utils";
 import { openai } from "@ai-sdk/openai";
 import { generateText } from "ai";
 
@@ -255,13 +255,11 @@ Extract the quote:`,
   } catch (error) {
     console.error("Error extracting quote with AI:", error);
     // Fallback to pattern-based extraction
-    return extractRelevantQuoteFallback(query, insight, content, type);
+    return extractRelevantQuoteFallback(content, type);
   }
 }
 
 function extractRelevantQuoteFallback(
-  query: string,
-  insight: string,
   content: string,
   _type: "win" | "risk" | "action",
 ): string {
@@ -306,30 +304,6 @@ function extractRelevantQuoteFallback(
   return "Key moment from the conversation";
 }
 
-function truncateQuote(quote: string): string {
-  // Clean up the quote
-  quote = quote.trim();
-
-  // If quote is reasonable length, return as is
-  if (quote.length <= 100) {
-    return quote;
-  }
-
-  // If too long, truncate at sentence boundary
-  const sentences = quote.split(/[.!?]+/);
-  if (sentences.length > 1 && sentences[0].length <= 80) {
-    return sentences[0].trim() + ".";
-  }
-
-  // Otherwise, truncate at word boundary
-  const words = quote.split(" ");
-  if (words.length > 20) {
-    return words.slice(0, 20).join(" ") + "...";
-  }
-
-  return quote;
-}
-
 async function generateCoachingTipWithAI(
   insight: string,
   type: string,
@@ -362,53 +336,6 @@ Coaching tip:`,
       return "Set clear expectations and follow through consistently to build trust and momentum.";
     }
   }
-}
-
-function generateTimestamp(): string {
-  // Generate a random timestamp within the conversation (0-60 minutes)
-  // More realistic distribution - most sales calls are 15-45 minutes
-  const minutes = Math.floor(Math.random() * 45) +
-    Math.floor(Math.random() * 15);
-  const seconds = Math.floor(Math.random() * 60);
-  return `${minutes.toString().padStart(2, "0")}:${
-    seconds.toString().padStart(2, "0")
-  }`;
-}
-
-function extractTimestampFromContent(content: string): string {
-  // Try to extract timestamp patterns from content
-  // Common formats: [00:15:30], (15:30), 15:30, etc.
-  const timestampPatterns = [
-    /\[(\d{1,2}:\d{2}:\d{2})\]/g, // [00:15:30]
-    /\[(\d{1,2}:\d{2})\]/g, // [15:30]
-    /\((\d{1,2}:\d{2}:\d{2})\)/g, // (00:15:30)
-    /\((\d{1,2}:\d{2})\)/g, // (15:30)
-    /(?:^|\s)(\d{1,2}:\d{2}:\d{2})(?:\s|$)/g, // 00:15:30
-    /(?:^|\s)(\d{1,2}:\d{2})(?:\s|$)/g, // 15:30
-  ];
-
-  for (const pattern of timestampPatterns) {
-    const matches = content.match(pattern);
-    if (matches && matches.length > 0) {
-      // Extract the timestamp from the first match
-      const match = matches[0];
-      const timestampMatch = match.match(/(\d{1,2}:\d{2}(?::\d{2})?)/);
-      if (timestampMatch) {
-        const timestamp = timestampMatch[1];
-        // Ensure format is MM:SS
-        if (timestamp.includes(":") && timestamp.split(":").length === 2) {
-          return timestamp;
-        } else if (timestamp.split(":").length === 3) {
-          // Convert HH:MM:SS to MM:SS by taking MM:SS part
-          const parts = timestamp.split(":");
-          return `${parts[1]}:${parts[2]}`;
-        }
-      }
-    }
-  }
-
-  // If no timestamp found in content, generate based on content characteristics
-  return generateTimestamp();
 }
 
 export async function POST(request: NextRequest) {
